Added PPM P6 file reader
Added convertion example (from P6 to P3)
Changed Pixel::Byte to be std::uint16_t, instead of std::uint8_t

+++ Pixel/Pixel.hpp
@@ -6,7 +6,7 @@
 class Pixel
 {
 public:
-    using Byte = std::uint8_t;
+    using Byte = std::uint16_t;
     using ColorValues = std::tuple<Byte, Byte, Byte>;
 
 public:
